Separate state management from presentation.
- Move all farm state, i/o & logic to composable useFarms.
- Move all farm UI to FarmList component.
