+++ src/farms.ts
@@ -1,4 +1,6 @@
-import { StatusObject } from "./status";
+import { invoke } from '@tauri-apps/api/tauri';
+import { onMounted, reactive } from 'vue';
+import { StatusObject, toStatusObject, RED, YELLOW, GREEN } from './status';
 
 export interface BackendFarmObject {
   id: number,
@@ -13,5 +15,52 @@ export interface FarmListItem {
   timestamp: number,
 }
 
-export const toBackendFarmObject = (f: FarmListItem): BackendFarmObject =>
+const toBackendFarmObject = (f: FarmListItem): BackendFarmObject =>
   ({ ...f, status: f.status.title });
+
+export default function useFarms() {
+  const farms: FarmListItem[] = reactive([]);
+  onMounted(() => {
+    invoke('load').then((cache: BackendFarmObject[]): void => {
+      const list: FarmListItem[] = cache.map((f) => ({
+        ...f, status: toStatusObject(f.status),
+      }));
+      farms.push(...list);
+    });
+  })
+
+  function save() {
+    const cache: BackendFarmObject[] = farms.map(toBackendFarmObject);
+    invoke('save', { farms: cache });
+  }
+
+  function setFarmStatus(i: number, color: string|symbol): void {
+    farms[i].status = toStatusObject(color);
+  }
+
+
+  const examples: FarmListItem[] = [
+    {
+      id: 0,
+      name: 'Joe\'s Farm',
+      status: toStatusObject(RED),
+      timestamp: 20
+    },
+    {
+      id: 1,
+      name: 'Sally\'s Farm',
+      status: toStatusObject(YELLOW),
+      timestamp: 5
+    },
+    {
+      id: 2,
+      name: 'Sam\'s Farm',
+      status: toStatusObject(GREEN),
+      timestamp: 0
+    },
+  ];
+
+  return {
+    examples, farms, save, setFarmStatus,
+  }
+}
